text2: replacing tests
